docs(yard): fix tests
