+++ documentation-site/components/yard/ast.ts
@@ -1,0 +1,64 @@
+import generate from '@babel/generator';
+import traverse from '@babel/traverse';
+import * as t from '@babel/types';
+import {parse} from 'react-view';
+
+export function toggleOverrideSharedProps(code: string, sharedProps: string[]) {
+  let result: string = '';
+  try {
+    const ast = parse(code) as any;
+    traverse(ast, {
+      ArrowFunctionExpression(path) {
+        if (result !== '') return;
+        if (path.node.params.length !== 1) return;
+        const firstParam: any = path.node.params[0];
+        let newParams: string[] = [];
+        if (firstParam.type === 'ObjectPattern') {
+          const properties = firstParam.properties;
+          newParams = properties.map((prop: any) => prop.key.name);
+        }
+
+        const shoudlWeAddSharedProps = newParams.every(
+          name => !sharedProps.includes(name),
+        );
+
+        if (shoudlWeAddSharedProps) {
+          sharedProps.forEach(param => {
+            if (!newParams.includes(param)) {
+              newParams.push(param);
+            }
+          });
+          path.node.params = [
+            //@ts-ignore
+            t.objectPattern(
+              newParams.map(param =>
+                t.objectProperty(
+                  t.identifier(param),
+                  t.identifier(param),
+                  false,
+                  true,
+                ),
+              ),
+            ),
+          ];
+        } else {
+          path.node.params = [
+            //@ts-ignore
+            t.objectPattern([
+              t.objectProperty(
+                t.identifier('$theme'),
+                t.identifier('$theme'),
+                false,
+                true,
+              ),
+            ]),
+          ];
+        }
+        result = generate(path.node as any).code;
+      },
+    });
+  } catch (e) {
+    throw new Error('Override params transform was no good.');
+  }
+  return result;
+}
