+++ documentation-site/components/yard/override.tsx
@@ -1,74 +1,12 @@
 import * as React from 'react';
-import generate from '@babel/generator';
-import traverse from '@babel/traverse';
-import * as t from '@babel/types';
 import {useStyletron} from 'baseui';
 import {StatefulTooltip} from 'baseui/tooltip';
 import {Button, KIND, SIZE} from 'baseui/button';
 import {ButtonGroup} from 'baseui/button-group';
 import Editor from './editor';
+import {toggleOverrideSharedProps} from './ast';
 
-import {formatCode, parse} from 'react-view';
-
-export function toggleOverrideSharedProps(code: string, sharedProps: string[]) {
-  let result: string = '';
-  try {
-    const ast = parse(code) as any;
-    traverse(ast, {
-      ArrowFunctionExpression(path) {
-        if (result !== '') return;
-        if (path.node.params.length !== 1) return;
-        const firstParam: any = path.node.params[0];
-        let newParams: string[] = [];
-        if (firstParam.type === 'ObjectPattern') {
-          const properties = firstParam.properties;
-          newParams = properties.map((prop: any) => prop.key.name);
-        }
-
-        const shoudlWeAddSharedProps = newParams.every(
-          name => !sharedProps.includes(name),
-        );
-
-        if (shoudlWeAddSharedProps) {
-          sharedProps.forEach(param => {
-            if (!newParams.includes(param)) {
-              newParams.push(param);
-            }
-          });
-          path.node.params = [
-            //@ts-ignore
-            t.objectPattern(
-              newParams.map(param =>
-                t.objectProperty(
-                  t.identifier(param),
-                  t.identifier(param),
-                  false,
-                  true,
-                ),
-              ),
-            ),
-          ];
-        } else {
-          path.node.params = [
-            //@ts-ignore
-            t.objectPattern([
-              t.objectProperty(
-                t.identifier('$theme'),
-                t.identifier('$theme'),
-                false,
-                true,
-              ),
-            ]),
-          ];
-        }
-        result = generate(path.node as any).code;
-      },
-    });
-  } catch (e) {
-    throw new Error('Override params transform was no good.');
-  }
-  return result;
-}
+import {formatCode} from 'react-view';
 
 export const getHighlightStyles = (
   isLightTheme: boolean,
